[#14] refactor: converter 패키지명 generator 로 수정
