+++ src/main/java/com/mallang/notification/application/NotificationEventHandler.java
@@ -5,7 +5,7 @@ import com.mallang.comment.domain.CommentWrittenEvent;
 import com.mallang.common.domain.DomainEvent;
 import com.mallang.notification.domain.Notification;
 import com.mallang.notification.domain.NotificationRepository;
-import com.mallang.notification.domain.converter.NotificationGeneratorComposite;
+import com.mallang.notification.domain.generator.NotificationGeneratorComposite;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
