--- src/test/java/com/mallang/notification/domain/converter/CommentWrittenNotificationGenerateTest.java
@@ -1,5 +1,0 @@
-package com.mallang.notification.domain.converter;
-
-class CommentWrittenNotificationGenerateTest {
-
-}
